refactor: move StopCondition class to new stop_condition.py in stop_condition package
