+++ gmatpyplus/stop_condition/__init__.py
@@ -1,0 +1,1 @@
+from .stop_condition import *
